Fixed user online status
