+++ server/src/cloud/userOnlineStatus.ts
@@ -1,0 +1,14 @@
+export const initializeUserOnlineStatus = () => {
+  (Parse.Cloud as any).onLiveQueryEvent(({event, sessionToken}) => {
+    if (event === "ws_disconnect") {
+      const query = new Parse.Query<Parse.Object>("_Session");
+
+      query.equalTo("sessionToken", sessionToken);
+      query.first({useMasterKey: true}).then((session) => {
+        const user = session.get("user");
+        user.unset("boards");
+        user.save(null, {useMasterKey: true});
+      });
+    }
+  });
+};
